feat:리뷰평균API

+++ src/main/java/com/example/BE/actor/ActorService.java
@@ -31,7 +31,7 @@ public class ActorService {
     private final ActorRepository actorRepository;
     private final MovieActorRepository movieActorRepository;
 
-    public void saveActor(long id) throws MalformedURLException {
+    public void saveActor(int id) throws MalformedURLException {
 
         try{
             String result = "";
