+++ src/main/java/com/example/BE/movie/MovieRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovieRepository extends JpaRepository<MovieEntity, Long> {
+public interface MovieRepository extends JpaRepository<MovieEntity, Integer> {
 }
+
